// Front/src/app/feature/tabla-usuarios/shared/model/tabla.ts
export interface RegisterUser {
    failure:   boolean;
    code:      number;
    message:   null;
    body:      User;
    timestamp: string;
}

export interface User {
    id?:        string;
    name:      string;
    email:     string;
    password:  string;
    phones:    Phone[];
    created?:   string;
    modified?:  string;
    lastLogin?: string;
    token?:     string;
    isActive?:  boolean;
}

export interface Phone {
    id?:         number;
    number:     number;
    cityCode:   string;
    contryCode: string;
}

// Generated by https://quicktype.io

export interface ConsultaTodos {
    failure:   boolean;
    code:      number;
    message:   string;
    body:      User[];
    timestamp: string;
}

export interface Mensaje {
    mensaje: string,
    codigo: number
}

export interface Filtro {
    name?: string,
    email?: string,
    startDate?:string
    endDate?:string,
}